unplug_vf_passthrough: don't try to delete representor netdev
The representor is bound to the SR-IOV VF therefore it can't be
removed. Adding a flag in the delete_ovs_vif_port that allows to
skip deleting the netdev interface.

Closes-Bug: #1700726

+++ vif_plug_ovs/linux_net.py
@@ -88,10 +88,11 @@ def update_ovs_vif_port(dev, mtu=None, interface_type=None, timeout=None):
 
 
 @privsep.vif_plug.entrypoint
-def delete_ovs_vif_port(bridge, dev, timeout=None):
+def delete_ovs_vif_port(bridge, dev, timeout=None, delete_netdev=True):
     _ovs_vsctl(['--', '--if-exists', 'del-port', bridge, dev],
                timeout=timeout)
-    _delete_net_dev(dev)
+    if delete_netdev:
+        _delete_net_dev(dev)
 
 
 def device_exists(device):
